changed uint64 type to uint256 on time variables to can use block.timestamp operator

+++ contracts/LockedDeal.sol
@@ -10,7 +10,7 @@ contract LockedDeal is LockedPoolzData {
         Pool storage pool = AllPoolz[_PoolId];
         if(now < pool.StartTime) return 0;
         if(pool.FinishTime < now) return SafeMath.sub(pool.StartAmount, pool.DebitedAmount);
-        uint64 totalPoolDuration = pool.FinishTime - pool.StartTime;
+        uint256 totalPoolDuration = pool.FinishTime - pool.StartTime;
         uint256 timePassed = now - pool.StartTime;
         uint256 timePassedPermille = SafeMath.mul(timePassed, 1000);
         uint256 ratioPermille = SafeMath.div(timePassedPermille, totalPoolDuration);
